Remove duplicate confest
